added driver_location api and rate_ride api

+++ ride/models.py
@@ -20,8 +20,8 @@ class Ride(models.Model):
     fare = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
-    rider_rating = models.IntegerField(null=True, blank=True)
-    driver_rating = models.IntegerField(null=True, blank=True)
+    rider_rating = models.FloatField(null=True, blank=True)
+    driver_rating = models.FloatField(null=True, blank=True)
     rider_feedback = models.TextField(null=True, blank=True)
     driver_feedback = models.TextField(null=True, blank=True)
